+++ core/core-api/src/test/java/com/woopaca/taximate/core/domain/party/service/concurrent/ParticipationServiceConcurrencyTest.java
@@ -1,9 +1,10 @@
-package com.woopaca.taximate.core.domain.party.service;
+package com.woopaca.taximate.core.domain.party.service.concurrent;
 
 import com.woopaca.taximate.core.domain.auth.AuthenticatedUserHolder;
 import com.woopaca.taximate.core.domain.fixture.ParticipationFixtures;
 import com.woopaca.taximate.core.domain.fixture.PartyFixtures;
 import com.woopaca.taximate.core.domain.fixture.UserFixtures;
+import com.woopaca.taximate.core.domain.party.service.ParticipationService;
 import com.woopaca.taximate.core.domain.user.User;
 import com.woopaca.taximate.storage.db.core.entity.ParticipationEntity;
 import com.woopaca.taximate.storage.db.core.entity.PartyEntity;
@@ -13,23 +14,23 @@ import com.woopaca.taximate.storage.db.core.repository.PartyRepository;
 import com.woopaca.taximate.storage.db.core.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.stream.IntStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @Slf4j
-@SpringBootTest
-class ParticipationServiceTest {
+@DisplayName("ParticipationService 동시성 테스트")
+@SpringBootTest(webEnvironment = WebEnvironment.NONE)
+class ParticipationServiceConcurrencyTest extends ConcurrencyTest {
 
     @Autowired
     private ParticipationService participationService;
@@ -81,23 +82,10 @@ class ParticipationServiceTest {
 
             @Test
             void 최대_참여_인원을_초과하지_않아야_한다() throws InterruptedException {
-                ExecutorService executorService = Executors.newFixedThreadPool(CONCURRENT_PARTICIPATE_REQUESTS);
-                CountDownLatch countDownLatch = new CountDownLatch(CONCURRENT_PARTICIPATE_REQUESTS);
-
-                IntStream.rangeClosed(1, CONCURRENT_PARTICIPATE_REQUESTS)
-                        .forEach(i -> executorService.execute(() -> {
-                            try {
-                                setSecurityContextForUser(i);
-                                participationService.participateParty(partyEntity.getId());
-                            } catch (Exception e) {
-                                log.error("Error: ", e);
-                            } finally {
-                                countDownLatch.countDown();
-                            }
-                        }));
-
-                countDownLatch.await();
-                executorService.shutdown();
+                executeTasksInParallel(index -> {
+                    setSecurityContextForUser(index);
+                    participationService.participateParty(partyEntity.getId());
+                }, CONCURRENT_PARTICIPATE_REQUESTS);
 
                 List<ParticipationEntity> participationEntities = participationRepository
                         .findByPartyId(partyEntity.getId());
